test(api): add integration test coverage across the app (#62)
* test(api): add integration test coverage across api modules

Adds integration specs for bookmarks, explore, feed, imports, internal,
me, medias, recos, users/bookmarks, users/pinned, and the revenuecat
webhook, plus reusable test fixtures for playlists, sessions, persons,
explore maps, and import staging rows.

Fixes two bugs surfaced while writing the tests:
- buildJsonbObject() dropped computed SQL fragments (only kept raw
  Columns), so playlist_like feed items never got a role.
- MePushTokensService.set() never persisted deviceType, and
  RecoTargetsService's alreadySent flag ignored reco status, staying
  true forever after a reco was deleted.
- recosTrending's mediaId came back as a string from postgres, which
  silently broke cursor pagination past page 1; migration 0078 casts
  it to integer and preserves the view's existing indexes.

* fix(repo): run assets:generate before test targets

api:test failed in CI on the new ui-features spec: nothing previously
imported @libs/assets in a Jest-run file, so the missing dependency
wiring for its generated (gitignored) manifest went unnoticed. test
and test-integration now depend on ^generate so it's produced first.

* test(api): add integration/unit test coverage across playlists, tv-series, users, and remaining modules

Adds integration specs for playlists (items, likes, saves, members,
featured, add-targets, poster, guards, permissions), tv-series (logs,
seasons, episodes, images, reviews, playlists, sync), users (follow,
followers, following, recos, movies, tv-series, playlists), movies,
persons, search, and feed-persons, plus unit specs for guards,
interceptors, the cache service, cursor/sql/storage utils, and the
is-nullable decorator.

Fixes four bugs surfaced while writing the tests:
- PlaylistsService.duplicate() returned the pre-copy playlist row, so
  the response's itemsCount was always stale by however many items
  got copied.
- PlaylistItemsService's list endpoints accepted a `type` filter but
  never applied it to the query.
- TvSeriesLogsService.set() never ran the bookmark/reco completion
  cascade when a series log was created directly (only movies did).
- UserRecosService had no privacy gating at all, exposing another
  user's recos regardless of their profile visibility.

* test(notify): add unit/integration test coverage, fix notification bugs found along the way

Adds a test-integration target (real Postgres + a real I18nModule, so
translation interpolation is actually exercised) alongside unit specs
for FcmService, ApnsService and NotifyService. Covers all 18 job types
handled by NotifyProcessor plus the unhandled-job and error-propagation
paths.

Fixes three bugs surfaced while writing the tests:
- reco:completed passed the recipient's name under the wrong i18n arg
  key, so it silently rendered as an empty string in the push
  notification instead of the intended name.
- auth.delete_account_email's translation keys didn't exist in either
  locale file, so the email fell back to the raw dot-path key as its
  subject/body.
- reco:received never guarded against a missing sender (every other
  actor lookup in the file does), so a deleted/missing sender produced
  a broken notification with an undefined name instead of being
  silently skipped like everywhere else.

Also fixes tsconfig/jest config gaps (missing jsx, missing .tsx in the
transform patterns) that broke compiling the file once a test actually
exercised the .tsx email templates, and mocks @react-email/render's
render() around an internal dynamic import Jest's CJS runtime can't
execute without --experimental-vm-modules.

* test(worker): add unit/integration test coverage, fix silently swallowed sync failures

Adds a test-integration target (real Postgres, faked Typesense client)
alongside a unit spec for SearchProcessor covering job dispatch, the
unhandled-job path, and error propagation.

Fixes one bug surfaced while writing the tests: SearchService.syncPlaylist's
upsert branch caught Typesense errors and only logged them, unlike
syncUser's upsert (which lets errors propagate). Since SearchProcessor
relies on errors bubbling up to trigger a BullMQ retry, a failed playlist
sync was silently marked successful — no retry, and the search index
would go stale with no signal beyond a log line.

* test(libs): add test coverage across shared libraries, fix bugs found along the way

Adds unit tests for rules (validation regexes), env (schema
defaults/required fields/transforms), i18n (locale fallback, header
parsing, dictionaries), assets (asset URL builder), the shared
notify/worker BullMQ clients, notify-types (Zod DTO validation),
query-client's React Query cache-mutation helpers (against a real
QueryClient instance), api-js's RealtimeManager socket lifecycle and
playlist role predicates, and ui's cn() classname helper. Bootstraps a
jest/tsconfig test harness from scratch for notify-types and ui, which
previously had none.

Fixes three bugs surfaced while writing the tests:
- USER_RULES.BIO.REGEX used `.` instead of `[\s\S]`, so the API
  rejected any bio containing a newline, contradicting its own
  paired "no blank lines" check which only makes sense for
  multi-line input.
- getLocaleFromHeaders's "case-insensitive" header lookup re-lowercased
  the already-lowercase lookup key instead of the headers object's
  actual keys, so non-canonically-cased headers were never found; it
  also didn't strip `;q=` quality suffixes from the first
  Accept-Language entry, so a header like `fr-FR;q=0.9` fell through
  to the default locale.
- removeFromInfiniteCache always decremented the first page's
  total_results by 1 even when the removed item wasn't found on any
  page. Since this helper runs against every cached infinite-scroll
  query matching a filter, this drifted displayed result counts down
  over time on removals the loaded pages never actually contained.

* fix(repo): wire generate dependencies for api-js and fix notify's eager env crash

api-js's OpenAPI codegen target was named "codegen" instead of "generate",
so it never ran before api-js's own tests or query-client's tests (which
both need the gitignored src/__generated__ output) — breaking CI on a
clean checkout, which never had that folder to begin with. Renames it to
"generate" to match the convention nx.json's test/test-integration
targetDefaults already rely on (^generate), and adds an explicit
self-dependency on both api-js's and assets's own test targets, since the
caret only reaches a project's dependencies, not the project itself.

Also fixes fcm.service.spec.ts: it transitively imports env.ts (via
fcm.provider.ts) purely as an import side effect, and env.ts calls
process.exit(1) on an unsatisfied schema — passed locally by accident
(ambient env vars) but crashed the Jest worker in CI's clean environment.
Mocks ../../env like apns.service.spec.ts already does.

* fix(ci): add missing PREFECT/imports env vars to mock .env

generate-openapi target now runs before api-js/query-client tests
(see 0a3082c) and loads apiSchema, which requires PREFECT_API_URL,
PREFECT_API_AUTH_STRING, and API_INTERNAL_IMPORTS_SECRET — never
added to CI's mock .env since nothing exercised full env validation
before that fix.

* fix(ci): use syntactically valid PKCS8 dummy key for AUTH_APPLE_PRIVATE_KEY

createBetterAuth eagerly calls generateAppleClientSecret at NestJS
bootstrap time (auth.service.ts), which runs jose's importPKCS8 on
the raw string. The literal "mock" placeholder isn't a valid PEM,
so it threw as soon as generate-openapi actually booted AppModule
(now that 0a3082c wired it as a real dependency). This throwaway
EC key is not a secret — it's never used to talk to Apple, only
to satisfy PKCS8 parsing during CI.

+++ libs/db/src/lib/schemas/reco.ts
@@ -92,7 +92,11 @@ export const recoRelations = relations(reco, ({ one }) => ({
 export const recosTrending = pgMaterializedView('recos_trending').as((qb) => {
   return qb
     .select({
-      mediaId: sql<number>`COALESCE(${reco.movieId}, ${reco.tvSeriesId})`.as('media_id'),
+      // Cast to integer: COALESCE of two bigint columns comes back from postgres as a string
+      // (node-postgres doesn't parse int8), which silently broke cursor pagination — the
+      // encoded cursor's `id` failed the `z.number()` decode schema on any page past the first.
+      // TMDB ids stay well under the int4 range, so this is a safe narrowing.
+      mediaId: sql<number>`COALESCE(${reco.movieId}, ${reco.tvSeriesId})::integer`.as('media_id'),
       type: reco.type,
       recommendationCount: sql<number>`cast(count(*) as int)`.as('recommendation_count'),
       trendingScore: sql<number>`
